ui fix

+++ frontend/src/components/Homepage/Query.js
@@ -143,9 +143,9 @@ const Content = () => {
                         </Grid>     
                         <FormControl className={classes.formControl}>
                             <Select value={select} onChange={(e) => setSelect(e.target.value)}>
-                                <MenuItem value='0'>Enter one short cytB sequence</MenuItem>
+                                <MenuItem value='0'>Enter one long cytB sequence</MenuItem>
                                 <MenuItem value='1'>Enter three short cytB sequences</MenuItem>
-                                <MenuItem value='2'>Demo: Select one short cytB lion sequence in the database</MenuItem>
+                                <MenuItem value='2'>Demo: Select one long cytB lion sequence in the database</MenuItem>
                                 <MenuItem value='3'>Demo: Select three short cytB lion sequences in the database</MenuItem>
                             </Select>
                         </FormControl>
